% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToArchive.R
\name{copyToArchive}
\alias{copyToArchive}
\title{copyToArchive}
\usage{
copyToArchive(path, file, name, tags = "", replace = FALSE)
}
\arguments{
\item{path}{path to the archive}

\item{file}{path of the file to be copied}

\item{name}{name used to identify the file}

\item{tags}{string of information about the file}

\item{replace}{delete existing entries with the same name (TRUE) or
add to archive as a new version (FALSE). Defaults to FALSE}
}
\description{
\code{copyToArchive} adds a data object to an existing archive.
}
\details{
\code{copyToArchive} copies a file to the archive and adds to the index.

see help(createArchive) for details of other archiving functions and of the
structure of the index.
}
\examples{
copyToArchive(path = "C:/Project/archive",
              file = "C:/temp/myFile.rds",
              name = "New Results",
              tags = "some interesting findings",
              replace = TRUE)

}
