% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArchiveIndex.R
\name{readArchiveIndex}
\alias{readArchiveIndex}
\title{readArchiveIndex}
\usage{
readArchiveIndex(path = NULL)
}
\arguments{
\item{path}{path to the archive

see help(createArchive) for details of other archiving functions and of the
structure of the index.}
}
\description{
\code{readArchiveIndex} returns the index of an archive.
}
\details{
\code{readArchiveIndex} returns and the index of an archive as a data frame.
}
\examples{
readArchiveIndex('C:/MyStore')

}
