% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFromArchive.R
\name{removeFromArchive}
\alias{removeFromArchive}
\title{removeFromArchive}
\usage{
removeFromArchive(path, id)
}
\arguments{
\item{path}{path to the archive}

\item{id}{id number(s) of the file(s) to be deleted}
}
\description{
\code{removeFromArchive} deletes a file from the archive.
}
\details{
\code{removeFromArchive} deletes one or more files from an archive, adjusts the index
 and notes the deletion in the history file.

see \code{\link[=createArchive]{createArchive()}} for details of other archiving functions and of the
structure of the index.
}
\examples{
\dontrun{removeFromArchive('C:/MyStore', id=32 )}

}
