% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceTag.R
\name{replaceTag}
\alias{replaceTag}
\title{replaceTag}
\usage{
replaceTag(path, id, newtag)
}
\arguments{
\item{path}{path to the data archive}

\item{id}{id number(s) of the tag that is to be replaced}

\item{newtag}{the replacement tag}
}
\description{
\code{replaceTag} replaces the tag in a selected file.
}
\details{
\code{replaceTag} replaces the tag of a selected file(s) and
notes the change in the history file. If the new tag is not
a single string then it is collapsed into one using " , "
(space comma space) as the separator.
When the collapsed new tag begin "+" it is
added to the end of the existing tag.

see \code{\link[=createArchive]{createArchive()}} for details of other archiving functions and of the
structure of the index.
}
\examples{
\dontrun{replaceTag(path='MyStore', id=32, newtag="some tag")}

}
