% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceTags.R
\name{replaceTags}
\alias{replaceTags}
\title{replaceTags}
\usage{
replaceTags(path, id, newtags)
}
\arguments{
\item{path}{path to the data archive}

\item{id}{id number(s) of entries to be deleted}

\item{newtags}{new tags to replace the olds ones}
}
\description{
\code{replaceTags} deletes files from the archive.
}
\details{
\code{replaceTags} deletes file from an archive.

see help(createArchive) for details of other archiving functions and of the
structure of the index.
}
\examples{
replaceTags(path='MyStore', id=32, newtags="some new tags" )

}
