% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToArchive.R
\name{saveToArchive}
\alias{saveToArchive}
\title{saveToArchive}
\usage{
saveToArchive(x, file, name = "", tag = "", replace = FALSE)
}
\arguments{
\item{x}{R object to be save}

\item{file}{path+name of the file that is to be used to save the object
(path must point to an archive).
Currently support rds and csv files}

\item{name}{name used to identify the file in the index. Defaults to the name of the file to
be saved without its extension.}

\item{tag}{single string of information about the file}

\item{replace}{delete existing entries with the same name (TRUE) or
add to archive as a new version (FALSE). Defaults to FALSE}
}
\description{
\code{saveToArchive} save an R object to an existing archive.
}
\details{
\code{saveToArchive} saves an object to the archive, adds it to the index and
records the transaction in the history file.
The name used in the index need not be unique. Multiple files with the same
identification name are treated as versions of the same data. If replace=TRUE
all files with that index name are deleted before the new file is copied to
the archive.
If there is already a file in the archive with the same filename as the
new file, the new file is given a replacement filename composed of the old
filename plus a version number.

see \code{\link[=createArchive]{createArchive()}} for details of other archiving functions and of the
structure of the index.
}
\examples{
\dontrun{saveToArchive(myData,
              file    = "C:/archive/myFile.rds",
              name    = "New Results",
              tag     = "some interesting findings",
              replace = TRUE)
         }

}
