% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyToArchive.R
\name{copyToArchive}
\alias{copyToArchive}
\title{copyToArchive}
\usage{
copyToArchive(path, file, name, tag = "", replace = FALSE)
}
\arguments{
\item{path}{path to the archive}

\item{file}{path of the file that is to be copied}

\item{name}{name used to identify the file in the index}

\item{tag}{single string of information about the file}

\item{replace}{delete existing entries with the same name (TRUE) or
add to archive as a new version (FALSE). Defaults to FALSE}
}
\description{
\code{copyToArchive} copies a file to an existing archive.
}
\details{
\code{copyToArchive} copies a file to the archive, adds it to the index and
records the transaction in the history file.
The name used in the index need not be unique. Multiple files with the same
identification name are treated as versions of the same data. If replace=TRUE
all files with that index name are deleted before the new file is copied to
the archive.
If there is already a file in the archive with the same filename as the
new file, the new file is given a replacement filename composed of the old
filename plus a version number.

see \code{\link[=createArchive]{createArchive()}} for details of other archiving functions and of the
structure of the index.
}
\examples{
\dontrun{copyToArchive(path = "C:/Project/archive",
              file    = "C:/temp/myFile.rds",
              name    = "New Results",
              tag     = "some interesting findings",
              replace = TRUE)
         }

}
