% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArchive.R
\name{createArchive}
\alias{createArchive}
\title{createArchive}
\usage{
createArchive(path, destroy = FALSE)
}
\arguments{
\item{path}{path to the new archive}

\item{destroy}{if TRUE & path exists the contents of path are destroyed!! Defaults to FALSE}
}
\description{
\code{createArchive} creates an archive folder with an index.
}
\details{
\code{createArchive} creates an archive folder with an empty index and
a history file that automatically documents changes to the archive.

The archive folder may already exist, but if so, it should be empty. Setting clear = TRUE
will destroy the contents of the archive folder and allow a new archive to be created.

\itemize{
  \item copyToArchive() copies a file to the archive.
  \item removeFromArchive() removes a file from the archive and deletes it.
  \item checkArchive() looks for inconsistencies between the folder content and the index.
  \item replaceTag() replaces the old tag with new one.
}

The index contains
\itemize{
  \item id       ... an automatically assigned identification number: 1, 2, 3, etc.
  \item name     ... string identifier assigned by the user
  \item tag      ... a single string of tags
  \item filename ... filename used in the archive
  \item datetime ... date and time when saved
}
}
\examples{
createArchive('C:/myStore', destroy = TRUE)

}
