% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArchive.R
\name{createArchive}
\alias{createArchive}
\title{createArchive}
\usage{
createArchive(path)
}
\arguments{
\item{path}{path to this archive}
}
\description{
\code{createArchive} creates an archive folder with an index.
}
\details{
\code{createArchive} creates an archive folder with an empty index.

The folder may already exist, but it must be empty.

\itemize{
  \item copyToArchive() copies a file to the archive.
  \item moveToArchive() moves a file into the archive.
  \item dropFromArchive() removes a file from the archive and deletes it.
  \item readArchiveIndex() returns the index of an archive as a tibble
  \item editTags() enables editing tags.
}

The index contains
\itemize{
  \item id ... an identification number 1, 2, 3, etc.
  \item name ... a name given to the saved object
  \item tags ... single string of tags
  \item filename ... name of file within the archive
  \item datetime ... date and time when saved
}
}
\examples{
createArchive('C:/myStore')

}
