% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeFromArchive.R
\name{removeFromArchive}
\alias{removeFromArchive}
\title{removeFromArchive}
\usage{
removeFromArchive(path, id)
}
\arguments{
\item{path}{path to the data archive}

\item{id}{id number(s) of entries to be deleted}
}
\description{
\code{removeFromArchive} deletes files from the archive.
}
\details{
\code{removeFromArchive} deletes file from an archive.

see help(createArchive) for details of other archiving functions and of the
structure of the index.
}
\examples{
removeFromArchive(path='MyStore', id=32 )

}
