% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceTags.R
\name{replaceTags}
\alias{replaceTags}
\title{replaceTags}
\usage{
replaceTags(path, id, newtags)
}
\arguments{
\item{path}{path to the data archive}

\item{id}{id number of the file to be deleted}

\item{newtags}{the replacement tags}
}
\description{
\code{replaceTags} replaces the tags in a selected file.
}
\details{
\code{replaceTags} replaces the tags of a selected file and
note the change in the history file. When the new tags begin "+" they
are added to the end of the existing tags.

see \code{\link[=createArchive]{createArchive()}} for details of other archiving functions and of the
structure of the index.
}
\examples{
\dontrun{replaceTags(path='MyStore', id=32, newtags="some tags")}

}
